+++ Source/GateRimSG1/Goauld/CompProperties_GoauldForcedImplantation.cs
@@ -5,6 +5,12 @@ namespace GateRimSG1.Goauld
     public class CompProperties_GoauldForcedImplantation : CompProperties
     {
         public float minimumTargetAgeYears = 13f;
+        public GoauldSymbioteOrigin symbioteOrigin = GoauldSymbioteOrigin.Goauld;
+        public bool allowForcedImplantation = true;
+        public bool allowRitualImplantation = true;
+        public bool allowAutonomousHuntToggle = true;
+        public bool allowVoluntaryImplantation = false;
+        public float voluntaryImplantationRange = 12f;
         public bool autonomousHuntingEnabled = true;
         public int autonomousScanIntervalTicks = 60;
         public float autonomousSearchRadius = 35f;